requirements.txt
fastapi
uvicorn
pprintpp
requests
beautifulsoup4
pandas
gevent
psycopg2-binary
sqlalchemy
python-dotenv
